c++, contracts: Avoid duplicating wrapper functions.
In some cases, we need to revisit the wrapper emission code
several times, as passes through instantiation can add new
uses.  We must ensure that we accumulate to the wrapper list
rather than re-starting it on each pass or, in some cases,
we end up visiting the same wrapper generation twice.

This patch revisits the dependent code and has a few small
cleanups plus amendment of comments and whitespace.

TODO: maybe consider some kind of worklist approach in
place of the hash_map (but non-urgent).

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>