lto/113207 - make fld_type_variant more picky
The below adds additional verification to fld_type_variant that
there's only one variant matching fld_type_variant_equal_p on the
chain.  The PR shows that variants built with build_qualified_type
can be equal to others in that regard but not with regard to
what build_qualified_type does.

	PR lto/113207
	* ipa-free-lang-data.cc (fld_type_variant): Add extra checking.