[testsuite] tolerate later success [PR108357]
On leon3-elf and presumably on other targets, the test fails due to
differences in calling conventions and other reasons, that add extra
gimple stmts that prevent the expected optimization at the expected
point.  The optimization takes place anyway, just a little later, so
tolerate that.


for  gcc/testsuite/ChangeLog

	PR tree-optimization/108357
	* gcc.dg/tree-ssa/pr108357.c: Tolerate later optimization.